package builder: update pkg_resources working set when loading package.
Rebuilding the working set allows to find the distribution which of the package
built with the package builder.
When leaving the "loaded" package context manager, the working set is reset to
the original state.

This allows us to do things like pkg_resources.get_distribution while the package
is loaded.