Make it possible to install a ticking creator globally.
In order to let nested creators tick the clock too, the ticking creator
can be installed globally when used as context manager.

In order to support replacing the creator a chain is used internally,
allowing to register any amount of creator wrappers without the need to
monkey patch the create function.